requirements.txt
Django==2.2.6
django-debug-toolbar==3.1.1
djangorestframework==3.11.0
drf-writable-nested==0.6.2
drf-yasg==1.17.1
gunicorn==20.0.4
pytest-django==4.1.0
mysqlclient==2.0.1
pytest-cov==2.10.1